Reworks flow, generates, reads, and then packages images
